dws_sportista/backend/requirements.txt
fastapi==0.111.0
passlib==1.7.4
pydantic==2.7.1
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
